@@ -4,7 +4,7 @@
   "description": "Pack sprites for PixiJS using TexturePacker CLI.",
   "author": "Martijn Swart <opensource@studiokloek.nl> (https://studiokloek.nl)",
   "license": "MIT",
-  "main": "dist/index.js",
+  "exports": "./dist/index.js",
   "repository": "github:studiokloek/pixi-texturepacker",
   "bugs": {
     "url": "https://github.com/studiokloek/pixi-texturepacker/issues"
@@ -15,8 +15,9 @@
     "watch": "rollup -c -w",
     "release": "np"
   },
+  "type": "module",
   "engines": {
-    "node": ">=8.10.0"
+    "node": "^12.20.0 || ^14.13.1 || >=16.0.0"
   },
   "files": [
     "dist/*",
@@ -30,28 +31,28 @@
   },
   "devDependencies": {
     "@rollup/plugin-run": "^2.0.2",
-    "eslint": "^7.15.0",
+    "eslint": "^7.27.0",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-unicorn": "^23.0.0",
-    "np": "^7.0.0",
-    "rollup": "^2.35.1",
+    "eslint-plugin-unicorn": "^32.0.1",
+    "np": "^7.5.0",
+    "rollup": "^2.50.2",
     "rollup-plugin-terser": "^7.0.2",
     "source-map-support": "^0.5.19"
   },
   "dependencies": {
     "camelcase": "^6.2.0",
-    "chalk": "^4.1.0",
+    "chalk": "^4.1.1",
     "find-up": "^5.0.0",
-    "fs-extra": "^9.0.1",
+    "fs-extra": "^10.0.0",
     "get-value": "^3.0.1",
-    "globby": "^11.0.1",
-    "log-symbols": "^4.0.0",
-    "meow": "^8.0.0",
+    "globby": "^11.0.3",
+    "log-symbols": "^5.0.0",
+    "meow": "^10.0.0",
     "object.defaults": "^1.1.0",
-    "ora": "^5.1.0",
+    "ora": "^5.4.0",
     "pupa": "^2.1.1",
     "sane": "^4.1.0",
-    "set-value": "^3.0.2",
+    "set-value": "^4.0.0",
     "throttle-debounce": "^3.0.1",
     "uppercamelcase": "^3.0.0"
   }
