@@ -7,7 +7,7 @@ export default {
   input: 'src/index.js',
   output: {
     file: 'dist/index.js',
-    format: 'cjs',
+    format: 'es',
     sourcemap: isDevelopmentMode
   },
   plugins: [
